refactor: 불필요한 주석 제거 및 패키지 구조 변경
