+++ src/main/java/store/lastdance/dto/expense/AnalyzeExpenseRequestDTO.java
@@ -1,12 +1,8 @@
 package store.lastdance.dto.expense;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
 
 @Schema(description = "LLM 분석 응답")
 public record AnalyzeExpenseRequestDTO(
