+++ src/main/java/store/lastdance/service/expense/ExpenseV2QueryServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import store.lastdance.converter.ExpenseConverter;
+import store.lastdance.converter.expense.ExpenseConverter;
 import store.lastdance.domain.common.ImageFile;
 import store.lastdance.domain.expense.*;
 import store.lastdance.domain.group.Group;
@@ -58,9 +58,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         );
     }
 
-    /**
-     * 지출 조회
-     */
     @Override
     public ExpenseResponseDTO getExpenseById(UUID userId, Long expenseId) {
         User user = findUserById(userId);
@@ -69,7 +66,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 () -> new CustomException(ErrorCode.EXPENSE_NOT_FOUND)
         );
 
-        // 그룹 지출 - 정산 데이터 포함
         List<SplitDataDTO> splitData = null;
         if (expense.getExpenseType() == ExpenseType.GROUP) {
             splitData = getSplitData(expense);
@@ -78,9 +74,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         return expenseConverter.toResponseDTO(expense, splitData);
     }
 
-    /**
-     * 지출 정산 데이터 조회
-     */
     private List<SplitDataDTO> getSplitData(Expense expense) {
         return expenseSplitRepository.findByExpense(expense)
                 .stream()
@@ -88,9 +81,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .toList();
     }
 
-    /**
-     * 그룹 분담금 조회
-     */
     @Override
     public List<GroupShareExpenseResponseDTO> getGroupShareExpenses(UUID userId, ExpenseSearchDTO searchDTO) {
         log.info("=== getGroupShareExpenses 호출 ===");
@@ -98,7 +88,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
 
         User user = findUserById(userId);
 
-        // 1. SHARE 타입 지출들 조회
         List<Expense> shareExpenses = expenseRepository.findShareExpensesByUserAndMonth(user, searchDTO.year(), searchDTO.month());
         log.info("조회된 SHARE 지출 개수: {}", shareExpenses.size());
 
@@ -110,16 +99,13 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                     log.debug("SHARE 분담 금액: {}", shareExpense.getAmount());
                     log.debug("원본 지출 ID: {}", shareExpense.getOriginalExpense() != null ? shareExpense.getOriginalExpense().getExpenseId() : null);
 
-                    // 2. 원본 그룹 지출 조회 (더 많은 정보를 위해)
                     Expense originalExpense = shareExpense.getOriginalExpense();
 
-                    // 3. 분할 정보 조회 (원본 지출 기준)
                     List<SplitDataDTO> splitData = null;
                     if (originalExpense != null) {
                         splitData = getSplitData(originalExpense);
                     }
 
-                    // 4. 그룹 이름 조회
                     String groupName = shareExpense.getGroup() != null ?
                             shareExpense.getGroup().getGroupName() : "";
                     log.debug("그룹 이름: {}", groupName);
@@ -140,19 +126,14 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .toList();
     }
 
-    /**
-     * 영수증 조회
-     */
     @Override
     public String getReceiptImageUrl(Long expenseId, UUID userId) {
 
         User user = findUserById(userId);
-        // 권한 체크 포함해서 지출 조회
         Expense expense = expenseRepository.findByExpenseIdWithPermission(expenseId, user).orElseThrow(
                 () -> new CustomException(ErrorCode.EXPENSE_NOT_FOUND)
         );
 
-        // 영수증 파일 없으면 null 반환
         ImageFile receiptImageFile = expense.getReceiptImageFile();
         if (receiptImageFile == null) {
             return null;
@@ -161,35 +142,25 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         return imageService.generatePresignedUrl(receiptImageFile.getFileId());
     }
 
-    /**
-     * 개인 지출 월별 추이 조회
-     */
     @Override
     public MonthlyExpenseTrendResponseDTO getPersonalExpenseTrend(UUID userId, ExpenseSearchDTO searchDTO) {
         log.info("개인 지출 추이 조회: userId={}, year={}, month={}, months={}, category={}", userId, searchDTO.year(), searchDTO.month(), searchDTO.months(), searchDTO.category());
 
         User user = findUserById(userId);
 
-        // 날짜 범위 계산
         DateRange dateRange = calculateDateRange(searchDTO.year(), searchDTO.month(), searchDTO.months());
 
         ExpenseCategory categoryEnum = searchDTO.category() != null ? ExpenseCategory.valueOf(searchDTO.category().toUpperCase()) : null;
-        // Repository 조회
         List<Expense> expenses = expenseRepository.findPersonalExpensesByMonthRange(user, dateRange.startDate(), dateRange.endDate(), categoryEnum);
 
-        // 월별 그룹핑 및 응답
         return createTrendResponse(expenses, dateRange);
     }
 
-    /**
-     * 그룹 지출 월별 추이 조회
-     */
     @Override
     public MonthlyExpenseTrendResponseDTO getGroupExpenseTrend(UUID userId, UUID groupId, ExpenseSearchDTO searchDTO) {
         User user = findUserById(userId);
         Group group = findGroupById(groupId);
 
-        // 그룹 멤버인지 확인
         boolean isMember = groupMemberRepository.existsByGroupAndUser(group, user);
         if (!isMember) {
             throw new CustomException(ErrorCode.GROUP_MEMBER_NOT_FOUND);
@@ -198,20 +169,14 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         log.info("그룹 지출 추이 조회: groupId={}, year={}, month={}, months={}, category={}",
                 groupId, searchDTO.year(), searchDTO.month(), searchDTO.months(), searchDTO.category());
 
-        // 날짜 범위 계산
         DateRange dateRange = calculateDateRange(searchDTO.year(), searchDTO.month(), searchDTO.months());
 
         ExpenseCategory categoryEnum = searchDTO.category() != null ? ExpenseCategory.valueOf(searchDTO.category().toUpperCase()) : null;
-        // Repository에서 데이터 조회
         List<Expense> expenses = expenseRepository.findGroupExpensesByMonthRange(group, dateRange.startDate(), dateRange.endDate(), categoryEnum);
 
-        // 월별 그룹핑 및 응답 생성
         return createTrendResponse(expenses, dateRange);
     }
 
-    /**
-     * 날짜 범위 계산
-     */
     private DateRange calculateDateRange(int year, int month, int months) {
         LocalDate endDate = LocalDate.of(year, month, 1).with(TemporalAdjusters.lastDayOfMonth());
         LocalDate startDate = endDate.minusMonths(months - 1).with(TemporalAdjusters.firstDayOfMonth());
@@ -219,19 +184,12 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         return new DateRange(startDate, endDate);
     }
 
-    /**
-     * 월별 추이 응답 생성
-     */
     private MonthlyExpenseTrendResponseDTO createTrendResponse(List<Expense> expenses, DateRange dateRange) {
         Map<String, List<ExpenseResponseDTO>> monthlyData = createMonthlyGrouping(expenses, dateRange);
         return expenseConverter.toMonthlyTrendResponseDTO(monthlyData, dateRange.startDate, dateRange.endDate);
     }
 
-    /**
-     * 지출 데이터를 월별로 그룹핑
-     */
     private Map<String, List<ExpenseResponseDTO>> createMonthlyGrouping(List<Expense> expenses, DateRange dateRange) {
-        // N+1 문제 해결: 모든 분담 내역을 한 번에 조회
         List<Expense> groupExpenses = expenses.stream()
                 .filter(e -> e.getExpenseType() == ExpenseType.GROUP)
                 .toList();
@@ -249,26 +207,21 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                     ));
         }
 
-        // 지출 데이터 월별로 그룹핑
         final Map<Long, List<SplitDataDTO>> finalSplitsMap = splitsByExpenseId;
         Map<String, List<ExpenseResponseDTO>> monthlyData = expenses.stream()
                 .collect(Collectors.groupingBy(
                         expense -> expense.getExpenseDate().format(DateTimeFormatter.ofPattern("yyyy-MM")),
-                        LinkedHashMap::new,  // 순서 보장
+                        LinkedHashMap::new,
                         Collectors.mapping(
                                 expense -> expenseConverter.toResponseDTO(expense, finalSplitsMap.get(expense.getExpenseId())),
                                 Collectors.toList()
                         )
                 ));
 
-        // 데이터 없는 달 빈 배열로 추가
         fillEmptyMonths(monthlyData, dateRange.startDate, dateRange.endDate);
         return sortByMonthKey(monthlyData);
     }
 
-    /**
-     * 데이터 업는 달에 빈 배열 추가
-     */
     private void fillEmptyMonths(Map<String, List<ExpenseResponseDTO>> monthlyData, LocalDate startDate, LocalDate endDate) {
         LocalDate current = startDate;
         while (!current.isAfter(endDate)) {
@@ -278,9 +231,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         }
     }
 
-    /**
-     * 월별 키로 정렬
-     */
     private Map<String, List<ExpenseResponseDTO>> sortByMonthKey(Map<String, List<ExpenseResponseDTO>> monthlyData) {
         return monthlyData.entrySet().stream()
                 .sorted(Map.Entry.comparingByKey())
@@ -291,9 +241,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
     private record DateRange(LocalDate startDate, LocalDate endDate) {
     }
 
-    /**
-     * 그룹 분담 지출 통계 관련 메서드
-     */
     @Override
     public PageWithSummaryResponse<GroupShareExpenseResponseDTO> getGroupShareExpensesWithPaging(
             UUID userId, UUID groupId, ExpenseSearchDTO searchDTO, Pageable pageable
@@ -302,23 +249,18 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         Group group = findGroupById(groupId);
 
         ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
-        // 1. 필터링이 적용된 페이징 조회
         Page<Expense> shareExpensesPage = expenseRepository.findShareExpensesByGroupAndMonthWithPagingFiltered(
                 user, group, searchDTO.year(), searchDTO.month(),
                 categoryEnum, searchDTO.search(), pageable
         );
 
-        // 2. DTO 변환 (N+1 해결)
         List<GroupShareExpenseResponseDTO> pageContent = convertShareExpensesToDTOs(shareExpensesPage.getContent());
         Page<GroupShareExpenseResponseDTO> shareExpenseResponsePage = new PageImpl<>(pageContent, pageable, shareExpensesPage.getTotalElements());
 
-
-        // 3. 통계 계산을 위한 전체 데이터 조회
         if (!shareExpenseResponsePage.hasContent()) {
             return PageWithSummaryResponse.of(shareExpenseResponsePage, ExpenseSummary.empty());
         }
 
-        // 전체 데이터 조회 (페이징 없이, 필터링 적용)
         List<Expense> allShareExpenses = expenseRepository.findShareExpensesByGroupAndMonthWithPagingFiltered(
                 user, group, searchDTO.year(), searchDTO.month(),
                 categoryEnum, searchDTO.search(), Pageable.unpaged()
@@ -326,28 +268,22 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
 
         List<GroupShareExpenseResponseDTO> allShareExpensesDTOs = convertShareExpensesToDTOs(allShareExpenses);
 
-        // 4. 통계 정보 계산
         ExpenseSummary summary = calculateShareExpensesSummary(allShareExpensesDTOs);
 
         return PageWithSummaryResponse.of(shareExpenseResponsePage, summary);
     }
 
-    /**
-     * Expense 리스트를 GroupShareExpenseResponseDTO 리스트로 변환 (N+1 문제 해결)
-     */
     private List<GroupShareExpenseResponseDTO> convertShareExpensesToDTOs(List<Expense> shareExpenses) {
         if (shareExpenses.isEmpty()) {
             return Collections.emptyList();
         }
 
-        // 1. 모든 원본 지출 객체를 수집
         List<Expense> originalExpenses = shareExpenses.stream()
                 .map(Expense::getOriginalExpense)
                 .filter(Objects::nonNull)
-                .distinct() // 중복 제거
+                .distinct()
                 .toList();
 
-        // 2. 모든 원본 지출의 분담 내역을 한 번에 조회
         final Map<Long, List<SplitDataDTO>> splitsByOriginalExpenseId;
         if (!originalExpenses.isEmpty()) {
             splitsByOriginalExpenseId = expenseSplitRepository.findByExpenseIn(originalExpenses).stream()
@@ -362,7 +298,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
             splitsByOriginalExpenseId = Collections.emptyMap();
         }
 
-        // 3. DTO 변환
         return shareExpenses.stream()
                 .map(shareExpense -> {
                     Expense originalExpense = shareExpense.getOriginalExpense();
@@ -381,20 +316,15 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .toList();
     }
 
-    /**
-     * GroupShareExpenseResponseDTO 리스트로 분담 지출 통계 정보 계산
-     */
     private ExpenseSummary calculateShareExpensesSummary(List<GroupShareExpenseResponseDTO> expenses) {
         if (expenses.isEmpty()) {
             return ExpenseSummary.empty();
         }
 
-        // 전체 지출 금액 (원본 지출 기준)
         BigDecimal totalAmount = expenses.stream()
-                .map(GroupShareExpenseResponseDTO::amount)  // 원본 지출 총액
+                .map(GroupShareExpenseResponseDTO::amount)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
 
-        // 내 분담금 총합 (새로 추가)
         BigDecimal myTotalShareAmount = expenses.stream()
                 .map(GroupShareExpenseResponseDTO::myShareAmount)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
@@ -403,18 +333,16 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
         );
 
-        // 최대 지출 찾기 추가
         GroupShareExpenseResponseDTO maxExpense = expenses.stream()
                 .max(Comparator.comparing(GroupShareExpenseResponseDTO::amount))
                 .orElse(null);
 
         long totalCount = expenses.size();
-        long myShareCount = expenses.size();  // 내 분담금 건수 (현재는 동일)
+        long myShareCount = expenses.size();
 
-        // 카테고리별 통계 계산 (내 분담금 기준)
         Map<String, CategoryStats> categoryStats = calculateCategoryStats(
                 expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.myShareAmount())).toList(),
-                myTotalShareAmount  // 내 분담금 총합 기준으로 변경
+                myTotalShareAmount
         );
 
         return new ExpenseSummary(
@@ -435,30 +363,22 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
 
         User user = findUserById(userId);
 
-        // 1. 개인 지출과 분담 지출 DTO 리스트를 각각 생성
         List<CombinedExpenseResponseDTO> personalExpenseDTOs = fetchPersonalExpenseDTOs(user, searchDTO);
         List<CombinedExpenseResponseDTO> shareExpenseDTOs = fetchShareExpenseDTOs(user, searchDTO);
 
-        // 2. 두 리스트를 통합하고 정렬
         List<CombinedExpenseResponseDTO> allExpenses = new ArrayList<>();
         allExpenses.addAll(personalExpenseDTOs);
         allExpenses.addAll(shareExpenseDTOs);
         allExpenses.sort(Comparator.comparing(CombinedExpenseResponseDTO::date).reversed());
 
-        // 3. 통계 정보 계산
         ExpenseSummary summary = calculateExpenseSummary(allExpenses);
 
-        // 4. 수동 페이징 처리
         Page<CombinedExpenseResponseDTO> page = applyManualPaging(allExpenses, pageable);
 
         return PageWithSummaryResponse.of(page, summary);
     }
 
-    /**
-     * 사용자의 개인 지출 내역을 조회하여 DTO 리스트 변환
-     */
     private List<CombinedExpenseResponseDTO> fetchPersonalExpenseDTOs(User user, ExpenseSearchDTO searchDTO) {
-
         ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
         List<Expense> personalExpenses = expenseRepository.findPersonalExpensesForCombined(
                 user, searchDTO.year(), searchDTO.month(), categoryEnum, searchDTO.search(), Pageable.unpaged()
@@ -469,9 +389,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .collect(Collectors.toList());
     }
 
-    /**
-     * 사용자의 그룹 분담 지출 내역을 조회하여 DTO 리스트 변환 (N+1 해결)
-     */
     private List<CombinedExpenseResponseDTO> fetchShareExpenseDTOs(User user, ExpenseSearchDTO searchDTO) {
         ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
         List<Expense> shareExpenses = expenseRepository.findShareExpensesForCombined(
@@ -491,9 +408,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .collect(Collectors.toList());
     }
 
-    /**
-     * 리스트에 대해 수동으로 페이징 처리 적용
-     */
     private <T> Page<T> applyManualPaging(List<T> sourceList, Pageable pageable) {
         int start = (int) pageable.getOffset();
         if (start > sourceList.size()) {
@@ -504,18 +418,12 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         return new PageImpl<>(pageContent, pageable, sourceList.size());
     }
 
-    /**
-     * 사용자가 해당 그룹의 멤버인지 검증
-     */
     private void validateGroupMembership(User user, Group group) {
         if (!groupMemberRepository.existsByGroupAndUser(group, user)) {
             throw new CustomException(ErrorCode.GROUP_MEMBER_NOT_FOUND);
         }
     }
 
-    /**
-     * 카테고리 문자열을 ExpenseCategory Enum 변환
-     */
     private ExpenseCategory parseCategory(String category) {
         if (category == null || category.trim().isEmpty()) {
             return null;
@@ -523,14 +431,10 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         try {
             return ExpenseCategory.valueOf(category.toUpperCase());
         } catch (IllegalArgumentException e) {
-            // 또는 로깅 후 null 반환 등 정책에 맞게 처리
             throw new CustomException(ErrorCode.INVALID_CATEGORY);
         }
     }
 
-    /**
-     * 조건에 따라 그룹 지출 내역을 데이터베이스에서 페이징하여 조회
-     */
     private Page<Expense> fetchGroupExpenses(Group group, ExpenseSearchDTO searchDTO, Pageable pageable) {
         String search = searchDTO.search();
         ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
@@ -549,15 +453,11 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         }
     }
 
-    /**
-     * CombinedExpenseResponseDTO 리스트로 통계 정보 계산
-     */
     private ExpenseSummary calculateExpenseSummary(List<CombinedExpenseResponseDTO> expenses) {
         if (expenses.isEmpty()) {
             return ExpenseSummary.empty();
         }
 
-        // 기본 통계 계산
         BigDecimal totalAmount = expenses.stream()
                 .map(CombinedExpenseResponseDTO::myShareAmount)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
@@ -566,21 +466,17 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
         );
 
-        // 최대 지출 찾기 추가
         CombinedExpenseResponseDTO maxExpense = expenses.stream()
                 .max(Comparator.comparing(CombinedExpenseResponseDTO::myShareAmount))
                 .orElse(null);
 
         long totalCount = expenses.size();
 
-
-        // 카테고리별 통계 계산
         Map<String, CategoryStats> categoryStats = calculateCategoryStats(
                 expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.myShareAmount())).toList(),
                 totalAmount
         );
 
-        // 개인/통합 지출의 경우 totalAmount와 myTotalShareAmount가 동일
         return new ExpenseSummary(
                 totalAmount,
                 averageAmount,
@@ -594,9 +490,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         );
     }
 
-    /**
-     * 공통 카테고리별 통계 계산
-     */
     private Map<String, CategoryStats> calculateCategoryStats(List<ExpenseStatItem> items, BigDecimal totalAmount) {
         Map<String, List<ExpenseStatItem>> categoryGroups = items.stream()
                 .collect(Collectors.groupingBy(ExpenseStatItem::category));
@@ -623,9 +516,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 ));
     }
 
-    /**
-     * 통계 계산용 내부 record
-     */
     private record ExpenseStatItem(String category, BigDecimal amount) {
     }
 
@@ -636,35 +526,25 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         User user = findUserById(userId);
         Group group = findGroupById(groupId);
 
-        // 1. 사전 조건 검증
         validateGroupMembership(user, group);
 
-        // 2. 페이징된 데이터 조회 (기존 fetchGroupExpenses 메서드 활용)
         Page<Expense> groupExpensesPage = fetchGroupExpenses(group, searchDTO, pageable);
 
         if (!groupExpensesPage.hasContent()) {
             return PageWithSummaryResponse.of(Page.empty(pageable), ExpenseSummary.empty());
         }
 
-        // 3. Page<Expense>를 Page<ExpenseResponseDTO>로 변환
         List<ExpenseResponseDTO> pageContent = convertToExpenseResponseDTOs(groupExpensesPage.getContent());
         Page<ExpenseResponseDTO> expenseResponsePage = new PageImpl<>(pageContent, pageable, groupExpensesPage.getTotalElements());
 
-
-        // 4. 통계 계산을 위한 전체 데이터 조회 (기존 fetchGroupExpenses 활용)
         List<Expense> allExpensesForStats = fetchGroupExpenses(group, searchDTO, Pageable.unpaged()).getContent();
         List<ExpenseResponseDTO> allExpenseDTOsForStats = convertToExpenseResponseDTOs(allExpensesForStats);
 
-
-        // 5. 통계 정보 계산
         ExpenseSummary summary = calculateGroupExpensesSummary(allExpenseDTOsForStats);
 
         return PageWithSummaryResponse.of(expenseResponsePage, summary);
     }
 
-    /**
-     * 지출 목록을 DTO로 변환
-     */
     private List<ExpenseResponseDTO> convertToExpenseResponseDTOs(List<Expense> expenses) {
         return expenses.stream()
                 .map(expense -> {
@@ -674,9 +554,6 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 .collect(Collectors.toList());
     }
 
-    /**
-     * ExpenseResponseDTO 리스트로 그룹 지출 통계 정보 계산
-     */
     private ExpenseSummary calculateGroupExpensesSummary(List<ExpenseResponseDTO> expenses) {
         if (expenses.isEmpty()) {
             return ExpenseSummary.empty();
@@ -690,14 +567,12 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
                 BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
         );
 
-        // 최대 지출 찾기 추가
         ExpenseResponseDTO maxExpense = expenses.stream()
                 .max(Comparator.comparing(ExpenseResponseDTO::amount))
                 .orElse(null);
 
         long totalCount = expenses.size();
 
-        // 기존 calculateCategoryStats 메서드 활용
         Map<String, CategoryStats> categoryStats = calculateCategoryStats(
                 expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.amount())).toList(),
                 totalAmount
@@ -716,14 +591,8 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         );
     }
 
-
-    /**
-     * LLM 지출 분석 기록 조회
-     */
     @Override
     public PageWithSummaryResponse<ExpenseAnalysisHistoryDTO> getExpenseAnalysisHistory(UUID userId, Pageable pageable) {
-
-//        User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         User user = findUserById(userId);
 
         Page<ExpenseAnalysisHistory> historyPage = expenseAnalysisHistoryRepository.findByUser(user, pageable);
@@ -731,8 +600,5 @@ public class ExpenseV2QueryServiceImpl implements ExpenseV2QueryService {
         Page<ExpenseAnalysisHistoryDTO> historyDTOPage = historyPage.map(ExpenseAnalysisHistoryDTO::from);
 
         return PageWithSummaryResponse.of(historyDTOPage, ExpenseSummary.empty());
-
     }
-
-
 }
